ros2/rdbox/oidc/ui/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
python-keycloak >= 0.27.0
requests >= 2.25.1
urllib3 >= 1.26.5